+++ projects/js-packages/charts/src/components/bar-list-chart/index.tsx
@@ -1,0 +1,1 @@
+export { default as BarListChart } from './bar-list-chart';