+++ projects/js-packages/boost-score-api/CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.1] - 2025-06-04
+### Changed
+- Update package dependencies. [#43766]
+
 ## [1.0.0] - 2025-06-03
 ### Changed
 - Internal updates. [#43068]
@@ -287,6 +291,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Create package for the boost score bar API [#30781]
 
+[1.0.1]: https://github.com/Automattic/jetpack-boost-score-api/compare/v1.0.0...v1.0.1
 [1.0.0]: https://github.com/Automattic/jetpack-boost-score-api/compare/v0.1.67...v1.0.0
 [0.1.67]: https://github.com/Automattic/jetpack-boost-score-api/compare/v0.1.66...v0.1.67
 [0.1.66]: https://github.com/Automattic/jetpack-boost-score-api/compare/v0.1.65...v0.1.66